requirements.txt
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6"
anyio==3.5.0; python_full_version >= "3.6.2" and python_version >= "3.6"
apscheduler==3.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
asgiref==3.5.0; python_version >= "3.8"
async-timeout==4.0.2; python_version >= "3.6"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
beautifulsoup4==4.10.0; python_full_version > "3.0.0"
bs4==0.0.1
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
dj-database-url==0.5.0
django-dramatiq==0.10.0; python_version >= "3.5"
django-stubs-ext==0.4.0; python_version >= "3.7"
django-stubs==1.10.1; python_version >= "3.7"
django==4.0.3; python_version >= "3.8"
dramatiq==1.12.3; python_version >= "3.6"
feedparser==6.0.8; python_version >= "3.6"
frozenlist==1.3.0; python_version >= "3.7"
google-api-core==2.6.0; python_version >= "3.6"
google-api-python-client==2.39.0; python_version >= "3.6"
google-auth-httplib2==0.1.0; python_version >= "3.6"
google-auth-oauthlib==0.5.0; python_version >= "3.6"
google-auth==2.6.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
googleapis-common-protos==1.55.0; python_version >= "3.6"
h11==0.12.0; python_version >= "3.6"
h2==4.1.0; python_version >= "3.6" and python_full_version >= "3.6.1"
hpack==4.0.0; python_version >= "3.6" and python_full_version >= "3.6.1"
httpcore==0.14.7; python_version >= "3.6"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
httpx==0.22.0; python_version >= "3.6"
hyperframe==6.0.1; python_version >= "3.6" and python_full_version >= "3.6.1"
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.6"
lxml==4.8.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
multidict==6.0.2; python_version >= "3.7"
mypy-extensions==0.4.3; python_version >= "3.7"
mypy==0.942; python_version >= "3.7"
oauthlib==3.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
packaging==21.3; python_version >= "3.6"
pika==1.2.0
prometheus-client==0.13.1; python_version >= "3.6"
protobuf==3.19.4; python_version >= "3.6"
psycopg2-binary==2.9.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and python_full_version >= "3.6.0"
pyparsing==3.0.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
pytz==2021.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
redis==4.2.1; python_version >= "3.6"
requests-oauthlib==1.3.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
rfc3986==1.5.0; python_version >= "3.6"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
sgmllib3k==1.0.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
sniffio==1.2.0; python_full_version >= "3.6.2" and python_version >= "3.6"
soupsieve==2.3.1; python_version >= "3.6" and python_full_version > "3.0.0"
sqlparse==0.4.2; python_version >= "3.8"
tomli==2.0.1; python_version >= "3.7"
twitchapi==2.5.2
types-pytz==2021.3.6; python_version >= "3.7"
types-pyyaml==6.0.5; python_version >= "3.7"
typing-extensions==4.1.1; python_version >= "3.7"
tzdata==2021.5; sys_platform == "win32" and python_version >= "3.8" and (python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6" and platform_system == "Windows") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6")
tzlocal==4.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
websockets==10.2; python_version >= "3.7"
whitenoise==6.0.0; python_version >= "3.7"
wrapt==1.14.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
yarl==1.7.2; python_version >= "3.6"
django-environ==0.8.1
